Add a greeting dialog
A greeting dialog was added that is automatically shown when the user
arrives at the site for the first time. When closing the dialog a string
will be saved to local storage and the dialog will not be shown again as
long as the string exists in the storage.

If the user later wants to see the dialog again a hamburger button was
added to the site that opens the dialog again.