Improve greeting dialog logic
Previously the greeting dialog display logic was based on if the user
has already closed the dialog or not. This meant that every time when
the user closed the dialog an item was stored in the local storage and
when entering the site the existence of the storage item key was
checked.

Current implementation shows the dialog if a specific local storage key
is not found and then immediately stores an item with the key.